+++ src/main/java/com/kevin/tiertagger/config/TTConfigScreen.java
@@ -21,12 +21,12 @@ public class TTConfigScreen extends AbstractConfigScreen<TierTaggerConfig> {
         return new SimpleOption[] {
                 SimpleOption.ofBoolean("tiertagger.config.enabled", config.isEnabled(), config::setEnabled),
                 new SimpleOption<>("tiertagger.config.gamemode", SimpleOption.emptyTooltip(), SimpleOption.enumValueText(),
-                        new SimpleOption.PotentialValuesBasedCallbacks<>(Arrays.asList(GameMode.values()), Codec.INT.xmap(GameMode::byId, GameMode::getId)),
+                        new SimpleOption.PotentialValuesBasedCallbacks<>(Arrays.asList(TierTaggerConfig.GameMode.values()), Codec.INT.xmap(TierTaggerConfig.GameMode::byId, TierTaggerConfig.GameMode::getId)),
                         config.getGameMode(), config::setGameMode),
                 SimpleOption.ofBoolean("tiertagger.config.unranked", config.isShowUnranked(), config::setShowUnranked),
                 SimpleOption.ofBoolean("tiertagger.config.retired", config.isShowRetired(), config::setShowRetired),
                 new SimpleOption<>("tiertagger.config.statistic", SimpleOption.emptyTooltip(), SimpleOption.enumValueText(),
-                        new SimpleOption.PotentialValuesBasedCallbacks<>(Arrays.asList(Statistic.values()), Codec.INT.xmap(Statistic::byId, Statistic::getId)),
+                        new SimpleOption.PotentialValuesBasedCallbacks<>(Arrays.asList(TierTaggerConfig.Statistic.values()), Codec.INT.xmap(TierTaggerConfig.Statistic::byId, TierTaggerConfig.Statistic::getId)),
                         config.getShownStatistic(), config::setShownStatistic),
                 Ukutils.createButton("tiertagger.clear", s -> TierCache.clearCache()),
         };
